Refactor: move vod search use cases tests
Moved all search use case tests from `domain/vod/src/test/java/usecase`
to `domain/vod/src/test/java/usecases/search`.

The following test files were moved:
- AddRecentViewedUseCaseTest.kt
- ClearRecentViewedUseCaseTest.kt
- ClearSearchHistoryUseCaseTest.kt
- GetRecentViewedUseCaseTest.kt
- GetSearchHistoryUseCaseTest.kt
- RemoveSearchHistoryUseCaseTest.kt
- SearchActorsUseCaseTest.kt
- SearchMoviesUseCaseTest.kt
- SearchTvSeriesUseCaseTest.kt

Additionally, JUnit Jupiter dependencies were added to `domain/vod/build.gradle.kts`.
