+++ domain/vod/build.gradle.kts
@@ -15,6 +15,8 @@ dependencies {
     implementation(libs.kotlinx.datetime)
     implementation(libs.kotlinx.coroutines.core)
     testImplementation(libs.junit.jupiter.api)
+    testImplementation(libs.junit.jupiter)
+    testImplementation(libs.junit.jupiter)
     testRuntimeOnly(libs.junit.jupiter.engine)
     testImplementation(libs.mockk)
     testImplementation(libs.kotlinx.coroutines.test)
