+++ domain/vod/src/test/java/usecases/search/GetSearchHistoryUseCaseTest.kt
@@ -1,4 +1,4 @@
-package usecase
+package usecases.search
 
 import com.google.common.truth.Truth.assertThat
 import exceptions.RetrievingDataFailureException
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import repository.SearchHistoryRepository
-import usecase.search.SearchHistory
+import search.search.SearchHistory
 
 class GetSearchHistoryUseCaseTest {
     private var searchHistoryRepository: SearchHistoryRepository = mockk(relaxed = true)
@@ -28,7 +28,9 @@ class GetSearchHistoryUseCaseTest {
     fun `execute() should return history list when available`() =
         runTest {
             // Given
-            coEvery { searchHistoryRepository.getSearchHistory(any()) } returns flowOf(SearchHistoryList)
+            coEvery { searchHistoryRepository.getSearchHistory(any()) } returns flowOf(
+                SearchHistoryList
+            )
 
             // When
             val result = getSearchHistoryUseCase.execute().single()
